chore(opensearch): refactor into cdk-lib
